+++ android/src/main/kotlin/com/simform/audio_waveforms/AudioWaveformsPlugin.kt
@@ -10,6 +10,7 @@ import androidx.annotation.RequiresApi
 import io.flutter.embedding.engine.plugins.FlutterPlugin
 import io.flutter.embedding.engine.plugins.activity.ActivityAware
 import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding
+import io.flutter.plugin.common.EventChannel
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
@@ -19,22 +20,59 @@ import java.io.IOException
 import java.text.SimpleDateFormat
 import java.util.*
 
+
 /** AudioWaveformsPlugin */
 class AudioWaveformsPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
     private lateinit var channel: MethodChannel
     private var recorder: MediaRecorder? = null
     private var activity: Activity? = null
-    private lateinit var audioWaveMethodCall: AudioWaveformsMethodCall
+    private lateinit var audioWaveMethodCall: AudioRecorder
     private var path: String? = null
     private var encoder: Int = 0
     private var outputFormat: Int = 0
     private var sampleRate: Int = 16000
 
+    //Todo: bitrate
+    private lateinit var audioPlayer: AudioPlayer
+
+    object Constants {
+        const val initRecorder = "initRecorder"
+        const val startRecording = "startRecording"
+        const val stopRecording = "stopRecording"
+        const val pauseRecording = "pauseRecording"
+        const val resumeRecording = "resumeRecording"
+        const val getDecibel = "getDecibel"
+        const val checkPermission = "checkPermission"
+        const val path = "path"
+        const val LOG_TAG = "AudioWaveforms"
+        const val methodChannelName = "simform_audio_waveforms_plugin/methods"
+        const val enCoder = "enCoder"
+        const val sampleRate = "sampleRate"
+        const val fileNameFormat = "dd-MM-yy-hh-mm-ss"
+        const val preparePlayer = "preparePlayer"
+        const val startPlayer = "startPlayer"
+        const val stopPlayer = "stopPlayer"
+        const val pausePlayer = "pausePlayer"
+        const val seekTo = "seekTo"
+        const val progress = "progress"
+        const val setVolume = "setVolume"
+        const val volume = "volume"
+        const val getDuration = "getDuration"
+        const val durationType = "durationType"
+        const val durationEventChannel = "durationEventChannel"
+        const val seekToStart = "seekToStart"
+    }
 
     override fun onAttachedToEngine(@NonNull flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
         channel = MethodChannel(flutterPluginBinding.binaryMessenger, Constants.methodChannelName)
         channel.setMethodCallHandler(this)
-        audioWaveMethodCall = AudioWaveformsMethodCall()
+        audioWaveMethodCall = AudioRecorder()
+        audioPlayer = AudioPlayer()
+        val eventChannel = EventChannel(
+            flutterPluginBinding.binaryMessenger,
+            Constants.durationEventChannel
+        )
+        eventChannel.setStreamHandler(audioPlayer)
     }
 
     @RequiresApi(Build.VERSION_CODES.N)
@@ -56,6 +94,37 @@ class AudioWaveformsPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
             Constants.resumeRecording -> audioWaveMethodCall.resumeRecording(result, recorder)
             Constants.getDecibel -> audioWaveMethodCall.getDecibel(result, recorder)
             Constants.checkPermission -> audioWaveMethodCall.checkPermission(result, activity)
+            Constants.preparePlayer -> {
+                val audioPath = call.argument(Constants.path) as String?
+                val volume = call.argument(Constants.volume) as Double?
+                audioPlayer.preparePlayer(result, audioPath, volume?.toFloat())
+            }
+            Constants.startPlayer -> {
+                val seekToStart = call.argument(Constants.seekToStart) as Boolean?
+                audioPlayer.start(result, seekToStart ?: true)
+            }
+            Constants.stopPlayer -> audioPlayer.stop(result)
+            Constants.pausePlayer -> audioPlayer.pause(result)
+            Constants.seekTo -> {
+                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+                    val progress = call.argument(Constants.progress) as Int?
+                    audioPlayer.seekToPosition(result, progress?.toLong())
+                } else {
+                    Log.e(
+                        Constants.LOG_TAG,
+                        "Minimum android O is required for seekTo function to works"
+                    )
+                }
+            }
+            Constants.setVolume -> {
+                val volume = call.argument(Constants.volume) as Float?
+                audioPlayer.setVolume(volume, result)
+            }
+            Constants.getDuration -> {
+                val type =
+                    if ((call.argument(Constants.durationType) as Int?) == 0) DurationType.Current else DurationType.Max
+                audioPlayer.getDuration(result, type)
+            }
             else -> result.notImplemented()
         }
     }