+++ lib/src/controllers/recorder_controller.dart
@@ -3,9 +3,9 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 
 import '/src/base/utils.dart';
-import 'audio_wave_interface.dart';
+import '../base/audio_waveforms_interface.dart';
 
-class WaveController extends ChangeNotifier {
+class RecorderController extends ChangeNotifier {
   final List<double> _waveData = [];
 
   ///At which rate waveform needs to be updated
@@ -47,6 +47,8 @@ class WaveController extends ChangeNotifier {
   ///If we have microphone permission or not.
   bool get hasPermission => _hasPermission;
 
+  bool shouldClearLabels = false;
+
   ///Use this to check permission and starts recording.
   ///
   ///Can be called after pausing.
@@ -70,7 +72,7 @@ class WaveController extends ChangeNotifier {
           _recorderState = RecorderState.initialized;
         }
         if (_recorderState == RecorderState.paused && Platform.isAndroid) {
-          _isRecording = await AudioWaveInterface.instance.resume();
+          _isRecording = await AudioWaveformsInterface.instance.resume();
           if (_isRecording) {
             _startTimer();
             _recorderState = RecorderState.recording;
@@ -82,7 +84,7 @@ class WaveController extends ChangeNotifier {
         }
 
         if (_recorderState == RecorderState.initialized) {
-          _isRecording = await AudioWaveInterface.instance.record(
+          _isRecording = await AudioWaveformsInterface.instance.record(
               Platform.isIOS ? iosEncoder.index : androidEncoder.index,
               sampleRate,
               path);
@@ -103,7 +105,7 @@ class WaveController extends ChangeNotifier {
 
   ///This method is only required for Android platform
   Future<void> _initRecorder(String? path) async {
-    final initialized = await AudioWaveInterface.instance.initRecorder(
+    final initialized = await AudioWaveformsInterface.instance.initRecorder(
         path, androidEncoder.index, androidOutputFormat.index, sampleRate);
     if (initialized) {
       _recorderState = RecorderState.initialized;
@@ -119,7 +121,7 @@ class WaveController extends ChangeNotifier {
   ///
   /// This method is called during record().
   Future<bool> checkPermission() async {
-    final result = await AudioWaveInterface.instance.checkPermission();
+    final result = await AudioWaveformsInterface.instance.checkPermission();
     if (result) {
       _hasPermission = result;
     }
@@ -131,7 +133,7 @@ class WaveController extends ChangeNotifier {
   ///Can start recording again after pausing.
   Future<void> pause() async {
     if (_recorderState == RecorderState.recording) {
-      _isRecording = (await AudioWaveInterface.instance.pause()) ?? true;
+      _isRecording = (await AudioWaveformsInterface.instance.pause()) ?? true;
       if (_isRecording) {
         throw "Failed to pause recording";
       }
@@ -150,12 +152,12 @@ class WaveController extends ChangeNotifier {
   Future<String?> stop([bool callReset = true]) async {
     if (_recorderState == RecorderState.recording ||
         _recorderState == RecorderState.paused) {
-      final path = await AudioWaveInterface.instance.stop();
+      final path = await AudioWaveformsInterface.instance.stop();
       if (path != null) {
         _isRecording = false;
         _timer?.cancel();
         _recorderState = RecorderState.stopped;
-        if (callReset) _reset();
+        if (callReset) reset();
         return path;
       } else {
         throw "Failed stop recording";
@@ -166,15 +168,25 @@ class WaveController extends ChangeNotifier {
     return null;
   }
 
-  void _reset() {
+  ///Clears WaveData and labels from the list.
+  void reset() {
     refresh();
     _waveData.clear();
+    shouldClearLabels = true;
     notifyListeners();
+  }
+
+  ///sets [shouldClearLabels] flag to false
+  void revertClearlabelCall() {
+    WidgetsBinding.instance?.addPostFrameCallback((_) {
+      shouldClearLabels = false;
+      notifyListeners();
+    });
   }
 
   ///gets decibels from native
   Future<double?> _getDecibel() async =>
-      await AudioWaveInterface.instance.getDecibel();
+      await AudioWaveformsInterface.instance.getDecibel();
 
   ///gets decibel by every defined frequency
   void _startTimer() {
@@ -225,10 +237,8 @@ class WaveController extends ChangeNotifier {
   ///This function must be called to free [resources],
   ///it will also dispose the controller.
   void disposeFunc() async {
-    if (_timer != null) {
-      _timer!.cancel();
-    }
-    await stop();
+    if (_timer != null) _timer!.cancel();
+    if (recorderState != RecorderState.stopped) await stop();
     dispose();
   }
 }